Move logic to server.
